selfmap: make isGlobals more flexible
My system has libraries in /usr/lib64 - the previous `find("/lib/")`
would return false for libc, libpthread, and everyone else, and BSS
segments wouldn't be recorded.  I think we want to record every lib,
but exclude libdoubletake (let me know if I'm reasoning about this
incorrectly).  There is probably a better way to do this -- what about
getting the current IP (or maybe just function pointer to current
method), find out which VMA it is in, get the filename associated with
_that_, and exclude the BSS for that lib.  That would avoid hardcoding
`libdoubletake`, but is obviously a bit more complicated.