log: cleanup & move much into source/log.cpp
Two goals here: support selecting the log-level at runtime without
incurring much additional overhead, and simplify some of the code.

What we're left with is a macro that checks a global atomic
corresponding to the current DoubleTake log-level against the
log-level of the statement (is it a WARNING, INFO, or DEBUG message),
calling into an out-of-line vararg logf function w/ the current file +
line info if we should emit something.

TODO: INFO and DEBUG levels seem to be switched from normal
definitions :\